Add files via upload
This is the Arduino script I developed originally for my custom smart watch I made.  I ported the code so it can work natively on the Adafruit Clue board.  As of now it's only function is to display and keep track of time.  The time can be synchronized by connecting the Adafruit Clue dev board via bluetooth to a device that supports CTS (all iOS devices would work).  When you disconnect from bluetooth, it will still keep track of the time, so no need for a RTC.  Hope someone out there finds this messy code helpful.  I left in a lot of other code so you can add a 16 LED NeoPixel ring that emulates hands on a clock, as well as code for an I2C small OLED display. 
- Michael Lacock, 2020